% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General_methods.R
\name{adlist}
\alias{adlist}
\alias{adlist,RHermesExp-method}
\title{adlist}
\usage{
adlist(struct)

\S4method{adlist}{RHermesExp}(struct)
}
\arguments{
\item{struct}{An RHermesExp object}
}
\value{
The current data.frame of adducts in struct.
}
\description{
adlist
}
\examples{
\dontshow{struct <- readRDS(system.file("extdata", "exampleObject.rds",
                             package = "RHermes"))}
adlist(struct)
}
\seealso{
Other Getters: 
\code{\link{Cluster}()},
\code{\link{DB}()},
\code{\link{IL}()},
\code{\link{Ident}()},
\code{\link{MS2Data}()},
\code{\link{PL}()},
\code{\link{SOI}()}
}
\author{
Roger Gine
}
\concept{Getters}
