% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MS2_functions.R
\name{CliqueMSMS}
\alias{CliqueMSMS}
\title{CliqueMSMS}
\usage{
CliqueMSMS(
  MS2Exp,
  idx,
  contaminant = 173.5,
  delta = 0.1,
  BPPARAM = BiocParallel::SerialParam(),
  sstype = "regular"
)
}
\value{
List of formulas and a list of purified MSMS (list of dataframes,
one for each superspectra).
}
\description{
Purify continuous MSMS spectra detecting consistent mass signals
}
\details{
First it parses each IL entry annotation to store the annotated
formulas that correspond to each entry. Then, it reads the MSMS info and
finds consistent mass traces by sorting all data points by mz and using
a Centwave-like algorithm.
The derived 'pure' spectra are called superspectra. A single IL entry
can yield different superspectra. This function is NOT to be used by
itself. It forms part of the MSMS processing workflow.
}
