% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General_methods.R
\name{Cluster}
\alias{Cluster}
\alias{Cluster,RHermesExp-method}
\title{Cluster}
\usage{
Cluster(struct)

\S4method{Cluster}{RHermesExp}(struct)
}
\arguments{
\item{struct}{An RHermesExp object}
}
\value{
The BiocParallel backend object associated with struct
}
\description{
Cluster
}
\examples{
Cluster(RHermesExp())
}
\seealso{
Other Getters: 
\code{\link{DB}()},
\code{\link{IL}()},
\code{\link{Ident}()},
\code{\link{MS2Data}()},
\code{\link{PL}()},
\code{\link{SOI}()},
\code{\link{adlist}()}
}
\author{
Roger Gine
}
\concept{Getters}
