% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General_methods.R
\name{DB}
\alias{DB}
\alias{DB,RHermesExp-method}
\title{DB}
\usage{
DB(struct)

\S4method{DB}{RHermesExp}(struct)
}
\arguments{
\item{struct}{An RHermesExp object}
}
\value{
A data.frame with all molecular formulas in struct.
}
\description{
DB
}
\examples{
\dontshow{struct <- readRDS(system.file("extdata", "exampleObject.rds",
                             package = "RHermes"))}
DB(struct)
}
\seealso{
Other Getters: 
\code{\link{Cluster}()},
\code{\link{IL}()},
\code{\link{Ident}()},
\code{\link{MS2Data}()},
\code{\link{PL}()},
\code{\link{SOI}()},
\code{\link{adlist}()}
}
\author{
Roger Gine
}
\concept{Getters}
