% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MS2_methods.R
\name{exportIdent}
\alias{exportIdent}
\alias{exportIdent,RHermesExp,numeric,character-method}
\title{exportIdent}
\usage{
exportIdent(struct, id, file)

\S4method{exportIdent}{RHermesExp,numeric,character}(struct, id, file)
}
\arguments{
\item{struct}{An RHermesExp object}

\item{id}{The id of the RHermesMS2Exp object.}

\item{file}{Name assigned to the csv file.}
}
\value{
Returns nothing, but generates a csv file with the exported
 identifications
}
\description{
Export all MS2 identifications (hits) into a csv file.
}
\examples{
\dontshow{struct <- readRDS(system.file("extdata", "exampleObject.rds",
                             package = "RHermes"))}
if(FALSE){exportIdent(struct, 1, "identifications.csv")}
}
\author{
Roger Gine
}
