% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MS2_methods.R
\name{exportMGF}
\alias{exportMGF}
\title{exportMGF}
\usage{
exportMGF(struct, id, file, whichSpec = NA)
}
\arguments{
\item{struct}{RHermesExp object}

\item{id}{Index of the MS2Exp that you want to export}

\item{file}{Name of the output file, without the .mgf termination}

\item{whichSpec}{Which entries to export. Defaults to all of them.}
}
\value{
An .msf file
}
\description{
Exports the superspectra of a given MS2Exp into an
 .mgf format file
}
\examples{
  if(FALSE){
       exportMGF(struct, 1, "output")
   }
}
\seealso{
Other Exports: 
\code{\link{exportMSP}()},
\code{\link{exportSIRIUS}()},
\code{\link{exportmzML}()}
}
\concept{Exports}
