% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MS2_methods.R
\name{exportMSP}
\alias{exportMSP}
\title{exportMSP}
\usage{
exportMSP(struct, id, file, whichSpec = NA)
}
\arguments{
\item{struct}{RHermesExp object}

\item{id}{Index of the MS2Exp that you want to export}

\item{file}{Name of the output file, without the .msp termination}

\item{whichSpec}{Which entries to export. Defaults to all of them.}
}
\value{
An .msp file
}
\description{
Exports the superspectra of a given MS2Exp ID into an
.msp format file
}
\examples{
  if(FALSE){
       exportMSP(struct, 1, "output")
   }
}
\seealso{
Other Exports: 
\code{\link{exportMGF}()},
\code{\link{exportSIRIUS}()},
\code{\link{exportmzML}()}
}
\concept{Exports}
