% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MS2_methods.R
\name{exportmzML}
\alias{exportmzML}
\title{exportmzML}
\usage{
exportmzML(struct, id, fname, whichSpec = NA, collisionEnergy = 35)
}
\arguments{
\item{struct}{RHermesExp object}

\item{id}{Index of the MS2Exp that you want to export}

\item{fname}{Name of the output file, without the .mzML termination}

\item{whichSpec}{Which entries to export. Defaults to all of them.}

\item{collisionEnergy}{Numeric. Desired collision energy (in %NCE, eV, or any
other unit)}
}
\value{
An .mzML file
}
\description{
Exports the superspectra of a given MS2Exp ID into an
mzML format file. Warning! The retention time information is not
included in the mzML file.
}
\examples{
  if(FALSE){
       exportmzML(struct, 1, "output")
   }
}
\seealso{
Other Exports: 
\code{\link{exportMGF}()},
\code{\link{exportMSP}()},
\code{\link{exportSIRIUS}()}
}
\author{
Jordi Capellades
}
\concept{Exports}
