% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MS2_methods.R
\name{exportSIRIUS}
\alias{exportSIRIUS}
\title{exportSIRIUS}
\usage{
exportSIRIUS(struct, id, file, whichSpec = NA)
}
\arguments{
\item{struct}{RHermesExp object}

\item{id}{Index of the MS2Exp that you want to export}

\item{file}{Name of the output file, without the .ms termination}

\item{whichSpec}{Which entries to export. Defaults to all of them.}
}
\value{
Returns nothing, but generates an .ms file where specified.
}
\description{
Exports all processed Hermes spectra into a .ms format suitable
for SIRIUS analysis.
}
\examples{
  if(FALSE){
       exportmzML(struct, 1, "output")
   }
}
\seealso{
Other Exports: 
\code{\link{exportMGF}()},
\code{\link{exportMSP}()},
\code{\link{exportmzML}()}
}
\concept{Exports}
