% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General_methods.R
\name{Ident}
\alias{Ident}
\alias{Ident,RHermesExp,numeric-method}
\title{Ident}
\usage{
Ident(struct, id)

\S4method{Ident}{RHermesExp,numeric}(struct, id)
}
\arguments{
\item{struct}{An RHermesExp object}

\item{id}{The number corresponding to the inclusion list you want to access}
}
\value{
The identification data.frame with the RHermes_MS2Exp at position id.
}
\description{
Ident
}
\examples{
\dontshow{struct <- readRDS(system.file("extdata", "exampleObject.rds",
                             package = "RHermes"))}
Ident(struct, 1)
}
\seealso{
Other Getters: 
\code{\link{Cluster}()},
\code{\link{DB}()},
\code{\link{IL}()},
\code{\link{MS2Data}()},
\code{\link{PL}()},
\code{\link{SOI}()},
\code{\link{adlist}()}
}
\author{
Roger Gine
}
\concept{Getters}
