% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General_methods.R
\name{IL}
\alias{IL}
\alias{IL,RHermesExp,numeric-method}
\title{IL}
\usage{
IL(struct, id)

\S4method{IL}{RHermesExp,numeric}(struct, id)
}
\arguments{
\item{struct}{An RHermesExp object}

\item{id}{The number corresponding to the inclusion list you want to access}
}
\value{
The RHermes_IL object at position id.
}
\description{
IL
}
\examples{
\dontshow{struct <- readRDS(system.file("extdata", "exampleObject.rds",
                             package = "RHermes"))}
IL(myHermes, 1)
}
\seealso{
Other Getters: 
\code{\link{Cluster}()},
\code{\link{DB}()},
\code{\link{Ident}()},
\code{\link{MS2Data}()},
\code{\link{PL}()},
\code{\link{SOI}()},
\code{\link{adlist}()}
}
\author{
Roger Gine
}
\concept{Getters}
