% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R
\docType{class}
\name{ILParam-class}
\alias{ILParam-class}
\alias{ILParam}
\title{ILParam}
\value{
An ILParam object
}
\description{
The ILParam class contains all details regarding IL generation,
 such as m/z window for grouping close SOIs, whether to prioritize or
 restrict the inclusion list to a subset of adducts, etc.
}
\details{
It's a part of the RHermesIL object and is an input for generateIL().
}
\section{Slots}{

\describe{
\item{\code{filtermz}}{Deltamz used to group SOIs with similar RTs. Defaults to
+-0.5Da. It should be matched with the m/z isolation window of your MS2
method settings.}

\item{\code{filterrt}}{Delta RT used to group SOIs with similar RTs. Defaults to 10s.}

\item{\code{rtmargin}}{Similar to filterrt. Defaults to 5s.}

\item{\code{priorization}}{Whether to prioritze adducts or not. Options are "only" to
select only a set of adduct annotations, "full" to cover all SOIs
independently of the adduct annotations or "yes" to specify some adducts to
be prioritized in order.}

\item{\code{ad}}{The specified adducts. See priorization.}
}}

\examples{
ILParam()
}
\seealso{
RHermesIL generateIL

Other Params: 
\code{\link{ExpParam-class}},
\code{\link{SOIParam-class}}
}
\author{
Roger Gine
}
\concept{Params}
