% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General_methods.R
\name{MS2Data}
\alias{MS2Data}
\alias{MS2Data,RHermesExp,numeric-method}
\title{MS2Data}
\usage{
MS2Data(struct, id)

\S4method{MS2Data}{RHermesExp,numeric}(struct, id)
}
\arguments{
\item{struct}{An RHermesExp object}

\item{id}{The number corresponding to the inclusion list you want to access}
}
\value{
A list of all IL entries of the RHermes_IL object at position id with
their corresponding MS2 scan data
}
\description{
MS2Data
}
\examples{
\dontshow{struct <- readRDS(system.file("extdata", "exampleObject.rds",
                             package = "RHermes"))}
MS2Data(struct, 1)
}
\seealso{
Other Getters: 
\code{\link{Cluster}()},
\code{\link{DB}()},
\code{\link{IL}()},
\code{\link{Ident}()},
\code{\link{PL}()},
\code{\link{SOI}()},
\code{\link{adlist}()}
}
\author{
Roger Gine
}
\concept{Getters}
