% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MS2Proc.R
\name{MS2Proc}
\alias{MS2Proc}
\title{MS2Proc}
\usage{
MS2Proc(
  struct,
  id,
  MS2files,
  referenceDB = "D:/sp_MassBankEU_20200316_203615.RData",
  mincos = 0.6,
  plotting = FALSE,
  outpdf = "mirror.pdf",
  outcsv = "./results.csv"
)
}
\arguments{
\item{struct}{The RHermesExp object to add the processed data.}

\item{id}{Numeric, the id of the inclusion list to process.}

\item{MS2files}{Character vector of the MS2 files address.}

\item{referenceDB}{Character, where to find the MSMS database to query against.}

\item{mincos}{Numeric, between 0 and 1, the minimum spectral cosine between query and
reference spectrum for a match to be reported (see the paper for more detail on this).}

\item{plotting}{Logical, whether you want mirror plots to be generated.}

\item{outpdf}{Character, where you want to store the plots (see plotting).}

\item{outcsv}{Character, where to store a summary csv of the identifications.}
}
\value{
An RHermesExp object with the identifications set in the used MS2Exp slot.
}
\description{
MS2Proc processes the MSMS files generated for a given IL and performs
compound identification using an MSMS reference spectra database. Only the struct
id and MS2files parameters are mandatory.
}
\examples{
MS2Proc(myHermes, 1, c('C:/myFolder/File1.mzML', 'C:/myFolder/File2.mzML'))

}
