% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_functions.R
\name{plotCoverage}
\alias{plotCoverage}
\alias{plotCoverage,RHermesExp,numeric-method}
\title{plotCoverage}
\usage{
plotCoverage(struct, id)

\S4method{plotCoverage}{RHermesExp,numeric}(struct, id)
}
\arguments{
\item{struct}{An RHermesExp object}

\item{id}{Number of the PL to use in the plot}
}
\value{
A list of two interactive plot_ly objects
}
\description{
Plots a representation of the raw data points covered by the
annotations and the redundancy of those annotations (that is, how many times
does the same data point get annotated as two different things).
}
\examples{
\dontshow{struct <- readRDS(system.file("extdata", "exampleObject.rds",
                             package = "RHermes"))}
if(FALSE){plotCoverage(struct, 1)}
}
\seealso{
Other Plots: 
\code{\link{plotIL}()},
\code{\link{plotISF}()},
\code{\link{plotPL}()}
}
\author{
Roger Gine
}
\concept{Plots}
