% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_functions.R
\name{plotFidelity}
\alias{plotFidelity}
\alias{plotFidelity,RHermesExp,numeric,numeric-method}
\title{plotFidelity}
\usage{
plotFidelity(struct, id, entry, plot = TRUE)

\S4method{plotFidelity}{RHermesExp,numeric,numeric}(struct, id, entry, plot = TRUE)
}
\arguments{
\item{struct}{An RHermesExp object.}

\item{id}{Number of the SOI to plot.}

\item{entry}{The SOI entry to check.}

\item{plot}{Default to TRUE. The parameter is used for
consistency with the function internal use in filterSOI(). If
set to FALSE, returns some statistics about the isotopic
fidelity.}
}
\value{
An interactive plot_ly object. If plot set to FALSE,
 returns a list of isotopic fidelity metrics.
}
\description{
Plots the selected SOI isotopic profile and a
 comparison with its theoretical abundances calculated from
 the molecular formula.
}
\examples{
\dontshow{struct <- readRDS(system.file("extdata", "exampleObject.rds",
                             package = "RHermes"))}
p <- plotFidelity(struct, 1, 9)
}
\seealso{
Other plots: 
\code{\link{plotRawMS2}()},
\code{\link{plotSOI}()},
\code{\link{plotSS}()}
}
\author{
Roger Gine
}
\concept{plots}
