% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOI_functions.R
\name{plotISF}
\alias{plotISF}
\title{plotISF}
\usage{
plotISF(struct, id)
}
\arguments{
\item{struct}{An RHermesExp object}

\item{id}{Number of the SOI list to plot}
}
\value{
A visNetwork plot
}
\description{
Plots all annotated ISF relations between SOIs. NOTE: you have
to first run removeISF with the parameter justAnnotate set to TRUE, otherwise
the ISF SOIs are removed from the SOI list.
}
\examples{
if(FALSE){
    plotISF(struct, 1)
}
}
\seealso{
Other Plots: 
\code{\link{plotIL}()}
}
\author{
Roger Gine
}
\concept{Plots}
