% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_functions.R
\name{plotPL}
\alias{plotPL}
\alias{plotPL,RHermesExp,numeric,character-method}
\title{plotPL}
\usage{
plotPL(
  struct,
  id,
  formula,
  ads = NA,
  rtrange = c(0, 10000),
  dynamicaxis = TRUE
)

\S4method{plotPL}{RHermesExp,numeric,character}(
  struct,
  id,
  formula,
  ads = NA,
  rtrange = c(0, 10000),
  dynamicaxis = TRUE
)
}
\arguments{
\item{struct}{An RHermesExp object}

\item{id}{Number of the file to plot}

\item{formula}{Formula annotation to search for (eg. "C6H12O6")}

\item{ads}{Adducts to plot. Defaults to NA, which plots all of them by
default.}

\item{rtrange}{The retention time interval to plot, in seconds (eg.
c(0,1000)). Defaults to a 0-10000s interval, which will cover all points.}

\item{dynamicaxis}{Whether to use a fixed y scale for all adducts or to adapt
the scale according to each adduct intensity}
}
\value{
An interactive plot_ly object
}
\description{
Plots the raw data points anotated with a given formula and
  adducts.
}
\examples{
\dontshow{struct <- readRDS(system.file("extdata", "exampleObject.rds",
                             package = "RHermes"))}
p <- plotPL(struct, 1, "C5H11NO2", c("M+H", "M+Na", "M+K"), c(80, 120))
}
\seealso{
Other Plots: 
\code{\link{plotCoverage}()},
\code{\link{plotIL}()},
\code{\link{plotISF}()}
}
\author{
Roger Gine
}
\concept{Plots}
