% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_functions.R
\name{plotRawMS2}
\alias{plotRawMS2}
\alias{plotRawMS2,RHermesExp,numeric-method}
\title{plotRawMS2}
\usage{
plotRawMS2(struct, ms2id, entryid)

\S4method{plotRawMS2}{RHermesExp,numeric}(struct, ms2id, entryid)
}
\arguments{
\item{struct}{An RHermesExp object}

\item{ms2id}{Number of the MS2Exp object where the spectrum is}

\item{entryid}{Number of the inclusion list entry to plot}
}
\value{
A list of four objects: a plot of the fragments grouped by their mz, a
 plot of the fragments grouped by their shape similarity, the similarity
 network and a summary table of peaks found and the group they belong to.
}
\description{
Plots an all MS2 fragments distribution along the rt of a given
 inclusion list entry.
}
\examples{
\dontshow{struct <- readRDS(system.file("extdata", "exampleObject.rds",
                             package = "RHermes"))}
plotRawMS2(struct, 1, 2)
}
\seealso{
Other plots: 
\code{\link{plotFidelity}()},
\code{\link{plotPL}()},
\code{\link{plotSOI}()},
\code{\link{plotSS}()}
}
\author{
Roger Gine
}
\concept{plots}
