% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_functions.R
\name{plotSOI}
\alias{plotSOI}
\alias{plotSOI,RHermesExp,numeric,character-method}
\title{plotSOI}
\usage{
plotSOI(
  struct,
  id,
  formula,
  ads = NA,
  rtrange = c(0, 10000),
  dynamicaxis = TRUE,
  interactive = TRUE
)

\S4method{plotSOI}{RHermesExp,numeric,character}(
  struct,
  id,
  formula,
  ads = NA,
  rtrange = c(0, 10000),
  dynamicaxis = TRUE,
  interactive = TRUE
)
}
\arguments{
\item{struct}{An RHermesExp object}

\item{id}{Number of the SOI list to plot}

\item{formula}{Formula annotation to search for (eg. "C6H12O6")}

\item{ads}{Adducts to plot. Defaults to NA, which plots all of them by
default.}

\item{rtrange}{The retention time interval to plot, in seconds (eg.
c(0,1000)). Defaults to a 0-10000s interval, which will cover all points.}

\item{dynamicaxis}{Whether to use a fixed y scale for all adducts or to adapt
the scale according to each adduct intensity}

\item{interactive}{Whether to return a plotly object or a ggplot. Defaults to
TRUE (plotly).}
}
\value{
An interactive plot_ly object or a static ggplot, depending on the
 value of interactive
}
\description{
Plots the SOI data points, the non-SOI points with the same
 annotation and, if blank subtraction was performed, it also plots the blank
 data points.
}
\examples{
\dontshow{struct <- readRDS(system.file("extdata", "exampleObject.rds",
                             package = "RHermes"))}
p <- plotSOI(struct, 1, "C5H11NO2", c("M+H", "M+Na", "M+K"), c(80, 120))
}
\seealso{
Other plots: 
\code{\link{plotFidelity}()},
\code{\link{plotRawMS2}()},
\code{\link{plotSS}()}
}
\author{
Roger Gine
}
\concept{plots}
