% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_functions.R
\name{plotSS}
\alias{plotSS}
\alias{plotSS,RHermesExp,numeric,numeric-method}
\title{plotSS}
\usage{
plotSS(struct, ms2id, ssnumber)

\S4method{plotSS}{RHermesExp,numeric,numeric}(struct, ms2id, ssnumber)
}
\arguments{
\item{struct}{An RHermesExp object}

\item{ms2id}{Number of the MS2Exp object where the spectrum is}

\item{ssnumber}{Number of the spectrum to plot}
}
\value{
A ggplot object
}
\description{
Plots an Hermes-cleaned MS2 spectum (mz/int).
}
\examples{
\dontshow{struct <- readRDS(system.file("extdata", "exampleObject.rds",
                             package = "RHermes"))}
if(FALSE){plotSS(struct, 1, 1)}
}
\seealso{
Other plots: 
\code{\link{plotFidelity}()},
\code{\link{plotRawMS2}()},
\code{\link{plotSOI}()}
}
\author{
Roger Gine
}
\concept{plots}
