% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MS2_functions.R
\name{processMS2}
\alias{processMS2}
\title{processMS2}
\usage{
processMS2(
  struct,
  id,
  MS2files,
  sstype = "regular",
  useDB = FALSE,
  referenceDB = "D:/sp_MassBankEU_20200316_203615.RData",
  mincos = 0.8,
  minhits = 1,
  method = "cosine"
)
}
\arguments{
\item{struct}{The RHermesExp object to add the processed data.}

\item{id}{Numeric, the id of the inclusion list to process.}

\item{MS2files}{Character vector of the MS2 files address.}

\item{referenceDB}{Character, where to find the MSMS database to query
against.}

\item{mincos}{Numeric, between 0 and 1, the minimum spectral cosine
between query and reference spectrum for a match to be reported
(see the paper for more detail on this).}
}
\value{
An RHermesExp object with the identifications set in the used
MS2Exp slot.
}
\description{
processMS2 processes the MSMS files generated for a given
IL and performs compound identification using an MSMS reference spectra
database. Only the struct id and MS2files parameters are mandatory.
}
\examples{
if(FALSE){
processMS2(myHermes, 1, c('C:/myFolder/File1.mzML',
                         'C:/myFolder/File2.mzML'))
}

}
