% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General_methods.R
\name{readTime}
\alias{readTime}
\alias{readTime,RHermesExp-method}
\title{readTime}
\usage{
readTime(struct)

\S4method{readTime}{RHermesExp}(struct)
}
\arguments{
\item{struct}{The RHermesExp object you want to read the timestamps from.}
}
\value{
None, just prints the timestamps
}
\description{
Prints all timestamps of a given RHermesExp object. Useful to
keep track of all changes made to the object (added files, generated SOI
 lists, changed parameters, etc.)
}
\examples{
readTime(RHermesExp())
}
\author{
Roger Gine
}
