% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metadata_methods.R
\name{remAd}
\alias{remAd}
\title{remAd}
\usage{
remAd(struct, name)
}
\arguments{
\item{struct}{The RHermesExp object}

\item{name}{The names of the adducts to remove. If they aren't in the
RHermesExp will raise a warning but no rows will be substracted.}
}
\value{
An updated RHermesExp object
}
\description{
Remove adducts from an RHermesExp adduct list
}
\details{
Removes a row (or many) from the adduct list slot inside the
RHermesExp object
}
\examples{
if(FALSE){
 remAd(myHermes, 'M+H') #Just one
 remAd(myHermes, c('M+DMSO+H', 'M+2K-H', 'M+CH3OH+H')) #Or many at a time
}
}
