% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOI_methods.R
\name{removeSOI}
\alias{removeSOI}
\title{removeSOI}
\usage{
removeSOI(struct, idx)
}
\value{
An RHermesExp object without the SOI lists indicated by idx.
}
\description{
Removes SOI lists from an RHermesExp object
}
\section{Slots}{

\describe{
\item{\code{struct}}{The RHermesExp object}

\item{\code{idx}}{The SOI lists to remove (indexes)}
}}

\examples{
if(FALSE){
myHermes <- removeSOI(myHermes, 2) #Remove the 2nd SOI list in the object
}
}
