% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R
\docType{class}
\name{RHermesData-class}
\alias{RHermesData-class}
\alias{RHermesData}
\title{RHermesData}
\description{
The experimental data storage class. Holds all information of the
peaklists, SOI lists, inclusion lists, MS2 data and identifications.
}
\section{Slots}{

\describe{
\item{\code{PL}}{A list object that holds all RHermesPL objects
containing a PL each.
See [RHermes]{RHermesPL} for more info}

\item{\code{SOI}}{List that holds RHermesSOI objects with a SOI list
inside each.
See [RHermes]{RHermesSOI} for more info.}

\item{\code{MS2Exp}}{List to hold inclusion lists, MSMS experimental
data and compound identifications.}
}}

