% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RHermesGUI.R
\name{RHermesGUI}
\alias{RHermesGUI}
\title{RHermesGUI}
\usage{
RHermesGUI()
}
\value{
Nothing, just starts the GUI
}
\description{
Calling this function starts the GUI in a new window on your
default browser.
}
\examples{
if(FALSE){RHermesGUI()}
}
\author{
Roger Gine
}
