% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R
\docType{class}
\name{RHermesMeta-class}
\alias{RHermesMeta-class}
\alias{RHermesMeta}
\title{RHermesMeta}
\description{
The metadata storage class. It holds the experimental parameters,
 all the names of the processed files and the timestamps for all operations
 performed on the parental RHermesExp object
}
\section{Slots}{

\describe{
\item{\code{ExpParam}}{Contains all experimental information (ppm error, resolution,
polarity, etc.), as well as the formula and adduct databases used.}

\item{\code{filenames}}{All filenames of the processed files. This info is also
available for individual PL and SOI objects in their respective slot.}

\item{\code{timestamps}}{Timestamps generated when any operation was performed on the
parental object. You can easily check them with readTime().}
}}

\seealso{
ExpParam readTime setTime setCluster Cluster
}
\author{
Roger Gine
}
