% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R, R/MS2_methods.R
\docType{class}
\name{RHermesMS2Exp-class}
\alias{RHermesMS2Exp-class}
\alias{RHermesMS2Exp}
\alias{show,RHermesMS2Exp-method}
\title{RHermesMS2Exp}
\usage{
\S4method{show}{RHermesMS2Exp}(object)
}
\arguments{
\item{object}{An RHermesMS2Exp object}
}
\value{
An RHermesMS2Exp object
}
\description{
The RHermesMS2Exp contains global information about the MS2 steps
 of the workflow. Particularly, it contains the inclusion list to be used,
 the MS2Data that has been acquired according to the IL and a list of
 purified MS2 spectra and identifications against a reference MS2 database
 (if performed)
}
\section{Slots}{

\describe{
\item{\code{IL}}{An RHermesIL object}

\item{\code{MS2Data}}{A list that contains header and peaks of the MS2 data acquired
for each IL entry. It gets its information after running processMS2.}

\item{\code{Ident}}{A list of different relevant information about the deconvoluted
MS2 spectra generated by processMS2.}
}}

\examples{
RHermesMS2Exp() #Do not generate the object manually
}
\seealso{
processMS2 generateIL IL Ident

Other Containers: 
\code{\link{RHermesIL-class}},
\code{\link{RHermesPL-class}},
\code{\link{RHermesSOI-class}}
}
\author{
Roger Gine
}
\concept{Containers}
