% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R, R/MS1_methods.R
\docType{class}
\name{RHermesPL-class}
\alias{RHermesPL-class}
\alias{RHermesPL}
\alias{show,RHermesPL-method}
\title{RHermesPL}
\usage{
\S4method{show}{RHermesPL}(object)
}
\arguments{
\item{object}{An RHermesPL object}
}
\value{
An RHermesPL object
}
\description{
The RHermesPL class contains all the annotation details from a
 single MS1 mzML file, as well as some raw information that is useful for
 other steps of the workflow. RHermesPL objects are generated by processMS1
 and should not be declared manually.
}
\section{Slots}{

\describe{
\item{\code{peaklist}}{Data.table containing all annotated data points. It stores, rt
values, mz, intensities, ionic formulas and isotope annotations.}

\item{\code{header}}{Dataframe with headers from the mzML file. Extracted with
mzR::header() and filtered empty scans (if any).}

\item{\code{raw}}{Dataframe of raw data points from the mzML file. Extracted with
mzR::peaks() and filtered by the noise parameter from ExpParam.}

\item{\code{labelled}}{Logical. Whether all possible 13C isotopologues were searched
for in the data. Usually FALSE, only TRUE if you're dealing with 13C
labelled data. Should not be set by the user here. Instead, tell processMS1
using labelled = TRUE.}

\item{\code{filename}}{Filename of the original mzML file.}
}}

\examples{
\dontshow{struct <- readRDS(system.file("extdata", "exampleObject.rds",
                             package = "RHermes"))}
RHermesPL() #Do NOT generate a PL object directly
PL(struct, 1) #Access the object using the getter function.
}
\seealso{
processMS1 ExpParam

Other Containers: 
\code{\link{RHermesIL-class}},
\code{\link{RHermesMS2Exp-class}},
\code{\link{RHermesSOI-class}}
}
\author{
Roger Gine
}
\concept{Containers}
