% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R, R/SOI_methods.R
\docType{class}
\name{RHermesSOI-class}
\alias{RHermesSOI-class}
\alias{RHermesSOI}
\alias{show,RHermesSOI-method}
\title{RHermesSOI}
\usage{
\S4method{show}{RHermesSOI}(object)
}
\arguments{
\item{object}{A RHermesSOI object}
}
\value{
An RHermesSOI object
}
\description{
The RHermesSOI class contains the SOI list, some precalculated
 information for plotting (in functions such as plotSOI)
}
\section{Slots}{

\describe{
\item{\code{SOIList}}{The SOI list itself.}

\item{\code{PlotDF}}{A dataframe contaning all SOI data points with annotations, ready
to be plotted when needed.}

\item{\code{SOIParam}}{A SOIParam object containing all information about SOI generation}

\item{\code{filename}}{The filename of the mzML used to generate the SOI list.}
}}

\examples{
\dontshow{struct <- readRDS(system.file("extdata", "exampleObject.rds",
                             package = "RHermes"))}
RHermesSOI() #Do NOT generate a SOI object directly, use findSOI
SOI(struct, 1) #Access the object using the getter function.
}
\seealso{
findSOI plotSOI plotIso

Other Containers: 
\code{\link{RHermesIL-class}},
\code{\link{RHermesMS2Exp-class}},
\code{\link{RHermesPL-class}}
}
\author{
Roger Gine
}
\concept{Containers}
