% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R
\name{setCluster}
\alias{setCluster}
\title{setCluster}
\usage{
setCluster(struct = NULL, cl = NULL)
}
\arguments{
\item{struct}{An RHermesExp object.}

\item{cl}{A BiocParallel backend object (like SerialParam(), SnowParam(),
MulticoreParam(), etc.).}
}
\value{
An RHermesExp object with an updated BiocParallel backend
}
\description{
By default selects the most appropiate BiocParallel backend
  according to your OS and RAM. If used with arguments, it assigns a
  specified backend to an RHermesExp object.
}
\examples{
struct <- RHermesExp()
struct <- setCluster(struct, SerialParam())
}
\author{
Roger Gine
}
