% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metadata_methods.R
\name{setDB}
\alias{setDB}
\alias{setDB,RHermesExp-method}
\title{setDB}
\usage{
setDB(
  struct,
  db = "hmdb",
  adcharge = 1,
  admult = 1,
  DBfile = "",
  adductfile = "",
  adlist = NA,
  keggpath = ""
)

\S4method{setDB}{RHermesExp}(
  struct,
  db = "hmdb",
  adcharge = 1,
  admult = 1,
  DBfile = "",
  adductfile = "",
  adlist = NA,
  keggpath = ""
)
}
\arguments{
\item{struct}{The RHermesExp object to update}

\item{db}{A character defining which database to use. You can test a sample of
HMDB or Norman using "hmdb" and "norman", respectively. Set it to "custom"
to use your own database. You can also specify some kegg pathways with
"kegg_p" and then a list of pathway IDs in the keggpath parameter.}

\item{adcharge}{The maximum charge (in absolute value, so polarity does not
matter) the adducts in the list can have. Defaults to 1.}

\item{admult}{The maximum multiplicity (M, 2M, 3M and so on) the adducts can
have. Defaults to 1.}

\item{DBfile}{Address to where the database is located. Can be either a csv
or a xls/xlsx.}

\item{adductfile}{Address to where the adduct list is located. Can only be a
csv file and should adhere to EnviPat adduct list format (run data(adducts,
package = "enviPat") to see how it's like).}

\item{adlist}{Vector of adduct names to use. It is an alternative to setting
your own adduct file or using the default lists.}

\item{keggpath}{A list of KEGG pathway IDs}
}
\value{
An RHermesExp object with the formula and adduct database set.
}
\description{
It updates the RHermesExp object and sets the formula database
 and adduct list to use in the following steps of the workflow. The adduct
 list used is based on the
 \href{http://pubs.acs.org/doi/abs/10.1021/acs.analchem.5b00941}{EnviPat}
 package
}
\examples{
if(FALSE){
    myHermes <- setDB(myHermes, 'hmdb') #Adcharge and admult default to 1
    myHermes <- setDB(myHermes, 'norman', 2, 1) #Charge 2, multiplicity 1
}


}
