% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metadata_methods.R
\name{setExpParam}
\alias{setExpParam}
\alias{setExpParam,RHermesExp-method}
\title{setExpParam}
\usage{
setExpParam(struct, params = ExpParam(), template = character(0))

\S4method{setExpParam}{RHermesExp}(struct, params = ExpParam(), template = character(0))
}
\arguments{
\item{struct}{The RHermesExp object to modify.}

\item{params}{Optional if using template. An ExpParam object detailing a
custom set of parameters.}

\item{template}{Ignored if using params. A character describing the ID of
 the pre-made parameter set to use. Currently you can use: 'orbi-pos',
'orbi-neg', 'qtof-pos' and 'qtof-neg'. The parameters can be found,
 modified or even expanded in the csv at app/www/InstrumentTemplates.csv}
}
\value{
An RHermesExp object with the desired experimental parameters set.
}
\description{
Main interface to set the experimental parameters into a
 RHermesExp object. The user can either provide an ExpParam object
(partially or fully customized) or provide a template tag to use a pre-made
 parameter set.
}
\examples{
if(FALSE){
myHermes <- RHermesExp()

#Using template
myHermes <- setExpParam(myHermes, template = 'orbi-pos')

#Custom parameters
myHermes <- setExpParam(myHermes,
                        ExpParam(ppm = 10, res = 135000, ion = '-'))
}

}
