% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General_methods.R
\name{setTime}
\alias{setTime}
\alias{setTime,RHermesExp,character-method}
\title{setTime}
\usage{
setTime(struct, message)

\S4method{setTime}{RHermesExp,character}(struct, message)
}
\arguments{
\item{struct}{The RHermesExp object you want to add the timestamp to.}

\item{message}{A character string to append.}
}
\value{
A copy of struct with the updated timestamp list.
}
\description{
Adds a timestamp to the RHermesExp object with a specified
message. Useful for adding info to the object for reproducibility purposes.
}
\examples{
setTime(RHermesExp(), "Important info")
}
\author{
Roger Gine
}
