% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General_methods.R
\name{SOI}
\alias{SOI}
\alias{SOI,RHermesExp,numeric-method}
\title{SOI}
\usage{
SOI(struct, id)

\S4method{SOI}{RHermesExp,numeric}(struct, id)
}
\arguments{
\item{struct}{An RHermesExp object}

\item{id}{The number corresponding to the SOI list you want to access}
}
\value{
The RHermes_SOI object at position id.
}
\description{
SOI
}
\examples{
\dontshow{struct <- readRDS(system.file("extdata", "exampleObject.rds",
                             package = "RHermes"))}
SOI(struct, 1)
}
\seealso{
Other Getters: 
\code{\link{DB}()},
\code{\link{IL}()},
\code{\link{Ident}()},
\code{\link{MS2Data}()},
\code{\link{PL}()},
\code{\link{adlist}()}
}
\author{
Roger Gine
}
\concept{Getters}
