% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOI_methods.R
\name{SOIsim}
\alias{SOIsim}
\alias{SOIsim,RHermesExp,numeric-method}
\title{SOIsim}
\usage{
SOIsim(struct, id, subset = NA, mode = "all")

\S4method{SOIsim}{RHermesExp,numeric}(struct, id, subset = NA, mode = "all")
}
\arguments{
\item{struct}{The RHermesExp object.}

\item{id}{The SOI list to be used.}

\item{subset}{A subset of SOI list entries that you want to compare the
profile similarities with. Defaults to NA, which means all the entries.}

\item{mode}{Whether to compare the similarity with "all" other SOIs or only
with those determined by subset. Defaults to "all", any other value
restricts the comparison to the subset.}
}
\value{
A similarity matrix.
}
\description{
Calculates the elution profile similarity between SOIs.
}
\examples{
if(FALSE){
myHermes <- SOIsim(myHermes, 2) #Remove the 2nd SOI list in the object
}
}
\author{
Roger Gine
}
