% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MS2_methods.R
\name{ssNetwork}
\alias{ssNetwork}
\title{ssNetwork}
\usage{
ssNetwork(struct, id, ss)
}
\arguments{
\item{struct}{An RHermesExp object}

\item{id}{The id of the RHermesMS2Exp object.}

\item{ss}{The MS2 spectra from Ident() to be compared}
}
\value{
A square symmetrical similarity matrix
}
\description{
Generates a similarity matrix between different MS2 spectra.
}
\examples{
\dontshow{struct <- readRDS(system.file("extdata", "exampleObject.rds",
                             package = "RHermes"))}
ssNetwork(struct, 1, 1:7)
}
\author{
Roger Gine
}
