% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportMS2.R
\name{exportMGF}
\alias{exportMGF}
\title{exportMGF}
\usage{
exportMGF(struct, id, fname = "output")
}
\arguments{
\item{struct}{RHermesExp object}

\item{ID}{Index of the MS2Exp that you want to export}

\item{fname}{Name of the output file, without the .mgf termination}
}
\description{
Exports the superspectra of a given MS2Exp into an .mgf format file
}
\examples{
  if(FALSE){
       exportMGF(myHermes, 1, "output")
   }
}
