% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MS2_methods.R
\name{exportmzML}
\alias{exportmzML}
\title{exportmzML}
\usage{
exportmzML(struct, ID, fname, idx = NA)
}
\arguments{
\item{struct}{RHermesExp object}

\item{ID}{Index of the MS2Exp that you want to export}

\item{fname}{Name of the output file, without the .mzML termination}
}
\value{
An .mzML file
}
\description{
Exports the superspectra of a given MS2Exp ID into an
mzML format file. Warning! The retention time information is not
included in the mzML file.
}
\examples{
  if(FALSE){
       exportmzML(myHermes, 1, "output")
   }
}
