% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R
\docType{class}
\name{ILParam-class}
\alias{ILParam-class}
\alias{ILParam}
\title{ILParam}
\value{
An ILParam object
}
\description{
The ILParam class contains all details regarding IL generation,
 such as m/z window for grouping close SOIs, whether to prioritize or
 restrict the inclusion list to a subset of adducts, etc.
}
\details{
It's a part of the RHermesIL object and is an input for generateIL().
}
\examples{
ILParam()
}
\seealso{
RHermesIL generateIL

Other Params: 
\code{\link{ExpParam-class}},
\code{\link{SOIParam-class}}
}
\author{
Roger Gine
}
\concept{Params}
