% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MS1_functions.R
\name{processMS1}
\alias{processMS1}
\title{processMS1}
\usage{
processMS1(struct, files, labelled = FALSE)
}
\arguments{
\item{struct}{RHermesExp S4 object to update with the processed data.
Important: The objects needs to have the experimental parameters already 
set before calling this function.}

\item{files}{Character vector of all the paths to the files to be processed.}

\item{labelled}{Logical, whether to check for all 13C isotopic signals.
Defaults to FALSE}
}
\value{
An RHermesExp object with the processed PLs.
}
\description{
Main function to process the input MS1 mzML files. It
accesses the files with mzR and processes them with RHermes internal
functions. For each file a PL object is generated inside the
object_data_PL slot.
}
\examples{
processMS1(myHermes, 'C:/myFolder/myFile.mzML', FALSE) #For regular use
processMS1(myHermes, 'C:/myFolder/myFile.mzML', TRUE) #For labelled data
}
