% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R
\docType{class}
\name{RHermesExp-class}
\alias{RHermesExp-class}
\alias{RHermesExp}
\title{RHermesExp}
\description{
The main RHermes class: The RHermesExp. It is a container for
all generated information. All main RHermes functions use it and return an
updated version of the object.
}
\section{Slots}{

\describe{
\item{\code{metadata}}{Where all the complementary info is stored (experimental
parameters, timestamps, databases, etc.). See [RHermes]{RHermesMeta} for more
info.}

\item{\code{data}}{All experimental info is stored in data. It is divided into PL
(peaklist) SOI (scans of interest) and MS2Exp (for the IL, MS2 data and 
identifications). See [RHermes]{RHermesData} for more info.}
}}

\examples{
if(FALSE){
 myHermes <- RHermesExp() #Initializing empty object
}
}
