% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R, R/IL_methods.R
\docType{class}
\name{RHermesIL-class}
\alias{RHermesIL-class}
\alias{RHermesIL}
\alias{show,RHermesIL-method}
\title{RHermesIL}
\usage{
\S4method{show}{RHermesIL}(object)
}
\arguments{
\item{object}{An RHermesIL object}
}
\value{
An RHermesIL object
}
\description{
The RHermesIL contains an inclusion list (IL) to perform the MS2
 experiment with it. It also contains information about whether adduct
 annotations have been prioritized over others, restricted, etc.
}
\details{
Should not be created directly by the user. Instead, use generateIL.
}
\section{Slots}{

\describe{
\item{\code{IL}}{The inclusion list itself}

\item{\code{annotation}}{A list of the SOIs corresponding to each IL entry}

\item{\code{SOInum}}{Which SOI list was used to generate the IL}

\item{\code{ILParam}}{An ILParam object containing all IL generation parameters}
}}

\examples{
\dontshow{struct <- readRDS(system.file("extdata", "exampleObject.rds",
                             package = "RHermes"))}
RHermesIL() #Do NOT generate an IL object directly
IL(struct, 1) #Access the object using the getter function.

}
\seealso{
generateIL exportIL plotIL filterIL ILParam

Other Containers: 
\code{\link{RHermesMS2Exp-class}},
\code{\link{RHermesPL-class}},
\code{\link{RHermesSOI-class}}
}
\author{
Roger Gine
}
\concept{Containers}
