% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R, R/MS2_methods.R
\docType{class}
\name{RHermesMS2Exp-class}
\alias{RHermesMS2Exp-class}
\alias{RHermesMS2Exp}
\alias{show,RHermesMS2Exp-method}
\title{RHermesMS2Exp}
\usage{
\S4method{show}{RHermesMS2Exp}(object)
}
\arguments{
\item{object}{An RHermesMS2Exp object}
}
\value{
An RHermesMS2Exp object
}
\description{
The SOIParam class contains all details regarding SOI detection,
 such as the number of RT bins and their width, minimum data point intensity,
 usage of blank substraction, etc.
}
\examples{
RHermesMS2Exp() #Do not generate the object manually
}
\seealso{
processMS2

Other Containers: 
\code{\link{RHermesIL-class}},
\code{\link{RHermesPL-class}},
\code{\link{RHermesSOI-class}}
}
\author{
Roger Gine
}
\concept{Containers}
