% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R
\name{setCluster}
\alias{setCluster}
\title{setCluster}
\usage{
setCluster()
}
\value{
A BiocParallel backend
}
\description{
Returns the most appropriate BiocParallel backend for RHermes
  according to your OS and RAM. It is entirely optional, but recommended
  nonetheless.
}
\examples{
BiocParallel::register(setCluster()) #Registers the most appropriate backend
}
\author{
Roger Gine
}
